refactor: reorganize project structure and layers
Refatoração da arquitetura do projeto:

- Separação clara em camadas: domain, application, infrastructure, presentation
- Domain: modelos e enums
- Application: serviços e DTOs de domínio
- Infrastructure: persistência (Mongo), configuração e segurança (JWT)
- Presentation: controllers, DTOs de API e exceptions
- Resources organizados: scripts, templates de email e reports
