+++ src/main/java/com/maisprati/hub/domain/enums/AppointmentStatus.java
@@ -1,0 +1,8 @@
+package com.maisprati.hub.domain.enums;
+
+public enum AppointmentStatus {
+    SCHEDULED,
+    CONFIRMED,
+    COMPLETED,
+    CANCELLED
+}
