+++ src/main/java/com/maisprati/hub/domain/enums/PhaseStatus.java
@@ -1,0 +1,7 @@
+package com.maisprati.hub.domain.enums;
+
+public enum PhaseStatus {
+    TODO,
+    IN_PROGRESS,
+    DONE
+}
