+++ src/main/java/com/maisprati/hub/domain/enums/PostType.java
@@ -1,0 +1,8 @@
+package com.maisprati.hub.domain.enums;
+
+public enum PostType {
+    TEAM_FORMATION,
+    GENERAL,
+    HELP_REQUEST,
+    ANNOUNCEMENT
+}
