+++ src/main/java/com/maisprati/hub/domain/enums/TeamMemberRole.java
@@ -1,0 +1,7 @@
+package com.maisprati.hub.domain.enums;
+
+public enum TeamMemberRole {
+    LEADER,
+    SUBLEADER,
+    MEMBER
+}
