+++ src/main/java/com/maisprati/hub/domain/model/TeamMember.java
@@ -1,0 +1,22 @@
+package com.maisprati.hub.domain.model;
+
+import com.maisprati.hub.domain.enums.TeamMemberRole;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class TeamMember {
+
+    private String userId;
+    private TeamMemberRole role;
+    private String subLeaderType; // Frontend, Backend, etc.
+    private LocalDateTime joinedAt;
+    private Boolean isActive;
+}
