+++ src/main/java/com/maisprati/hub/infrastructure/security/jwt/JwtProperties.java
@@ -1,0 +1,17 @@
+package com.maisprati.hub.infrastructure.security.jwt;
+
+import lombok.Data;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Data
+@Component
+@ConfigurationProperties(prefix = "jwt")
+public class JwtProperties {
+	
+	/** Chave secreta do JWT */
+	private String secret;
+	
+	/** Expiração do token em segundos */
+	private long expirationSeconds;
+}
