--- src/main/java/com/maisprati/hub/model/Comment.java
@@ -1,34 +1,0 @@
-package com.maisprati.hub.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import jakarta.validation.constraints.NotBlank;
-import java.time.LocalDateTime;
-
-@Document(collection = "comments")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Comment {
-
-    @Id
-    private String id;
-
-    @NotBlank
-    private String postId;
-
-    @NotBlank
-    private String authorId;
-
-    @NotBlank
-    private String content;
-
-    private Boolean isActive;
-    private LocalDateTime createdAt;
-}
