--- src/main/java/com/maisprati/hub/model/Post.java
@@ -1,37 +1,0 @@
-package com.maisprati.hub.model;
-
-import com.maisprati.hub.model.enums.PostType;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import jakarta.validation.constraints.NotBlank;
-import java.time.LocalDateTime;
-
-@Document(collection = "posts")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Post {
-
-    @Id
-    private String id;
-
-    @NotBlank
-    private String authorId;
-
-    @NotBlank
-    private String title;
-
-    @NotBlank
-    private String content;
-
-    private PostType postType;
-    private Boolean isActive;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-}
