--- src/main/java/com/maisprati/hub/model/TimeSlot.java
@@ -1,17 +1,0 @@
-package com.maisprati.hub.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class TimeSlot {
-
-    private String time;
-    private Boolean available;
-    private Boolean booked;
-}
