--- src/main/java/com/maisprati/hub/model/enums/AppointmentStatus.java
@@ -1,8 +1,0 @@
-package com.maisprati.hub.model.enums;
-
-public enum AppointmentStatus {
-    SCHEDULED,
-    CONFIRMED,
-    COMPLETED,
-    CANCELLED
-}
