--- src/main/java/com/maisprati/hub/model/enums/PhaseStatus.java
@@ -1,7 +1,0 @@
-package com.maisprati.hub.model.enums;
-
-public enum PhaseStatus {
-    TODO,
-    IN_PROGRESS,
-    DONE
-}
