--- src/main/java/com/maisprati/hub/model/enums/PostType.java
@@ -1,8 +1,0 @@
-package com.maisprati.hub.model.enums;
-
-public enum PostType {
-    TEAM_FORMATION,
-    GENERAL,
-    HELP_REQUEST,
-    ANNOUNCEMENT
-}
