--- src/main/java/com/maisprati/hub/model/enums/TeamMemberRole.java
@@ -1,7 +1,0 @@
-package com.maisprati.hub.model.enums;
-
-public enum TeamMemberRole {
-    LEADER,
-    SUBLEADER,
-    MEMBER
-}
